requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
bcrypt==4.1.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
databases==0.8.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
